requirements.txt
django==5.0.3
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
django-filter==24.2
psycopg2-binary==2.9.9
python-dotenv==1.0.1
beautifulsoup4==4.14.3
requests==2.32.5
selenium==4.41.0
selenium-stealth==1.0.6
drf-spectacular==0.29.0
